work_it_out/requirements.txt
django
justfile
pillow
easy_thumbnails
djangorestframework_simplejwt
djangorestframework
django-cors-headers